+++ app/(standard)/layout.js
@@ -1,4 +1,4 @@
-import { Box, Button, ButtonGroup, Typography } from "@mui/material";
+import { Box } from "@mui/material";
 import TopBar from "../components/TopBar";
 import BottomBar from "../components/BottomBar";
 import Avatar from "../components/Avatar";
@@ -19,19 +19,6 @@ export default function StandardLayout({ children }) {
         >
           {children}
         </Box>
-        <section>
-          <Typography variant="h6" sx={{ mt: 5, mb: 1 }}>
-            Go To:
-          </Typography>
-          <ButtonGroup>
-            <Button href="/setup" variant="outlined">
-              Create Comic
-            </Button>
-            <Button href="/create" variant="outlined">
-              Drawing Canvas
-            </Button>
-          </ButtonGroup>
-        </section>
       </Box>
       <BottomBar />
     </>
